피드 댓글 조회 처음부터 전부 다 반환

+++ src/provider/feed-comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException } from '@nestjs/common';
 import { FeedCommentRepository } from 'src/repository/feed-comment.repository';
 import { AwsService } from './aws.service';
 import { getImageUrl } from './util/get-image-url';
@@ -11,6 +11,23 @@ export class FeedCommentService {
   ) {}
 
   async create(userId: string, input: CreateFeedCommentInput) {
+    const promiseMethods = [
+      this.feedCommentRepository.existsFeed(input.feedId),
+    ];
+
+    if (input.parentCommentId) {
+      promiseMethods.push(
+        this.feedCommentRepository.existsComment(input.parentCommentId),
+      );
+    }
+
+    const [feedExists, commentExists] = await Promise.all(promiseMethods);
+
+    if (!feedExists) throw new HttpException('FEED', 404);
+    if (input.parentCommentId && !commentExists) {
+      throw new HttpException('COMMENT', 404);
+    }
+
     await this.feedCommentRepository.create(userId, input);
 
     if (input.mentionedUserId && input.parentCommentId) {
@@ -37,7 +54,36 @@ export class FeedCommentService {
     return;
   }
 
-  async getAllByFeedId(userId: string | null, feedId: string) {
+  async getComments(userId: string | null, feedId: string) {
+    const comments = await this.feedCommentRepository.findManyByFeedId(
+      userId,
+      feedId,
+    );
+
+    return comments.map((comment) => ({
+      ...comment,
+      writer: {
+        ...comment.writer,
+        image: getImageUrl(comment.writer.image),
+      },
+      childComments: comment.childComments.map((childComment) => ({
+        ...childComment,
+        writer: {
+          ...childComment.writer,
+          image: getImageUrl(childComment.writer.image),
+        },
+        mentionedUser: childComment.mentionedUser
+          ? {
+              ...childComment.mentionedUser,
+              image: getImageUrl(childComment.mentionedUser.image),
+            }
+          : null,
+      })),
+    }));
+  }
+
+  // 삭제
+  async getAllParentsByFeedId(userId: string | null, feedId: string) {
     const [comments, commentCount] = await Promise.all([
       this.feedCommentRepository.findAllParentsByFeedId(userId, feedId),
       this.feedCommentRepository.countByFeedId(feedId),
@@ -55,6 +101,7 @@ export class FeedCommentService {
     };
   }
 
+  // 삭제
   async getChildComments(
     userId: string | null,
     input: {
@@ -88,11 +135,21 @@ export class FeedCommentService {
   }
 
   async like(userId: string, commentId: string) {
+    const commentExists =
+      await this.feedCommentRepository.existsComment(commentId);
+
+    if (!commentExists) throw new HttpException('COMMENT', 404);
+
     await this.feedCommentRepository.createLike(userId, commentId);
     return;
   }
 
   async unlike(userId: string, commentId: string) {
+    const commentExists =
+      await this.feedCommentRepository.existsComment(commentId);
+
+    if (!commentExists) throw new HttpException('COMMENT', 404);
+
     await this.feedCommentRepository.deleteLike(userId, commentId);
     return;
   }
